[dsp++] clang warnings

+++ dsp++/include/dsp++/snd/format.h
@@ -177,7 +177,7 @@ public:
 	bool is_channel_present(channel::type::label ch) const {return channel_layout_.test(ch);}
 	void set_channel_present(channel::type::label ch, bool present = true)
 	{channel_layout_.set(ch, present); channel_count_ = static_cast<unsigned>(channel_layout_.count());}
-	unsigned channel_config() const {return channel_layout_.to_ulong();}
+	unsigned long channel_config() const {return channel_layout_.to_ulong();}
 	//! @return index of specified channel in current layout or channel::not_present if missing.
 	unsigned channel_index(channel::type::label ch) const;
 	unsigned channel_count() const {return channel_count_;}
